Rename to IntentVerifyLoginLink
